+++ main.go
@@ -25,7 +25,7 @@ import (
 
 // version is overridable at build time via -ldflags "-X main.version=...".
 // The Makefile injects the value from the VERSION file / git.
-var version = "0.6.0"
+var version = "0.7.0"
 
 func main() {
 	log.SetFlags(0)